update tool tests.
