+++ tool/test/org/antlr/v4/test/tool/TestATNConstruction.java
@@ -36,6 +36,8 @@ import org.antlr.v4.automata.ParserATNFactory;
 import org.antlr.v4.parse.ANTLRParser;
 import org.antlr.v4.runtime.atn.ATN;
 import org.antlr.v4.runtime.atn.ATNState;
+import org.antlr.v4.test.AntlrTestcase;
+import org.antlr.v4.test.ErrorQueue;
 import org.antlr.v4.tool.ErrorType;
 import org.antlr.v4.tool.Grammar;
 import org.antlr.v4.tool.LexerGrammar;
@@ -52,7 +54,7 @@ import java.util.Map;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-public class TestATNConstruction extends BaseTest {
+public class TestATNConstruction extends AntlrTestcase {
 	@Test
 	public void testA() throws Exception {
 		Grammar g = new Grammar(
