+++ tool/test/org/antlr/v4/test/tool/TestActionSplitter.java
@@ -34,6 +34,7 @@ import org.antlr.runtime.ANTLRStringStream;
 import org.antlr.runtime.Token;
 import org.antlr.v4.parse.ActionSplitter;
 import org.antlr.v4.semantics.BlankActionSplitterListener;
+import org.antlr.v4.test.AntlrTestcase;
 import org.junit.Test;
 
 import java.util.ArrayList;
@@ -41,7 +42,7 @@ import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 
-public class TestActionSplitter extends BaseTest {
+public class TestActionSplitter extends AntlrTestcase {
     static String[] exprs = {
         "foo",		"['foo'<" + ActionSplitter.TEXT + ">]",
         "$x",		"['$x'<" + ActionSplitter.ATTR + ">]",
