+++ tool/test/org/antlr/v4/test/tool/TestActionTranslation.java
@@ -30,11 +30,12 @@
 
 package org.antlr.v4.test.tool;
 
+import org.antlr.v4.test.AntlrTestcase;
 import org.junit.Test;
 
 /** */
 @SuppressWarnings("unused")
-public class TestActionTranslation extends BaseTest {
+public class TestActionTranslation extends AntlrTestcase {
 	String attributeTemplate =
 		"attributeTemplate(members,init,inline,finally,inline2) ::= <<\n" +
 		"parser grammar A;\n"+
