+++ tool/test/org/antlr/v4/test/tool/TestBufferedTokenStream.java
@@ -30,19 +30,14 @@
 
 package org.antlr.v4.test.tool;
 
-import org.antlr.v4.runtime.ANTLRInputStream;
-import org.antlr.v4.runtime.BufferedTokenStream;
-import org.antlr.v4.runtime.CharStream;
-import org.antlr.v4.runtime.LexerInterpreter;
-import org.antlr.v4.runtime.Token;
-import org.antlr.v4.runtime.TokenSource;
-import org.antlr.v4.runtime.TokenStream;
+import org.antlr.v4.runtime.*;
+import org.antlr.v4.test.AntlrTestcase;
 import org.antlr.v4.tool.LexerGrammar;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class TestBufferedTokenStream extends BaseTest {
+public class TestBufferedTokenStream extends AntlrTestcase {
 
 	protected TokenStream createTokenStream(TokenSource src) {
 		return new BufferedTokenStream(src);
