+++ tool/test/org/antlr/v4/test/tool/TestDollarParser.java
@@ -1,12 +1,12 @@
 package org.antlr.v4.test.tool;
 
-import org.antlr.v4.test.rt.java.BaseTest;
+import org.antlr.v4.test.AntlrTestcase;
 import org.junit.Test;
 
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
-public class TestDollarParser extends BaseTest {
+public class TestDollarParser extends AntlrTestcase {
 
 	@Test
 	public void testSimpleCall() throws Exception {
@@ -15,8 +15,8 @@ public class TestDollarParser extends BaseTest {
 	                  "  ;\n" +
 	                  "ID : 'a'..'z'+ ;\n";
 		String found = execParser("T.g4", grammar, "TParser", "TLexer", "a", "x", true);
-		assertTrue(found.indexOf(this.getClass().getSimpleName())>=0);
-		assertNull(this.stderrDuringParse);
+		assertTrue(found.contains(this.getClass().getSimpleName()));
+		assertNull(stderrDuringParse());
 	}
 
 }
