+++ tool/test/org/antlr/v4/test/tool/TestErrorSets.java
@@ -29,12 +29,13 @@
  */
 package org.antlr.v4.test.tool;
 
+import org.antlr.v4.test.AntlrTestcase;
 import org.antlr.v4.tool.ErrorType;
 import org.junit.Test;
 
 
 /** Test errors with the set stuff in lexer and parser */
-public class TestErrorSets extends BaseTest {
+public class TestErrorSets extends AntlrTestcase {
 	protected boolean debug = false;
 
 	/** Public default constructor used by TestRig */
