+++ tool/test/org/antlr/v4/test/tool/TestParserExec.java
@@ -30,6 +30,7 @@
 
 package org.antlr.v4.test.tool;
 
+import org.antlr.v4.test.AntlrTestcase;
 import org.junit.Ignore;
 import org.junit.Test;
 
@@ -67,7 +68,7 @@ import static org.junit.Assert.assertNull;
  *  Nongreedy loops match as much input as possible while still allowing
  *  the remaining input to match.
  */
-public class TestParserExec extends BaseTest {
+public class TestParserExec extends AntlrTestcase {
 
 
 	/**
@@ -92,7 +93,7 @@ public class TestParserExec extends BaseTest {
 			"s1-INT->s2\n" +
 			"s2-EOF->:s3=>1\n"; // Must point at accept state
 		assertEquals(expecting, result);
-		assertNull(this.stderrDuringParse);
+		assertNull(stderrDuringParse());
 	}
 
 	/**
@@ -102,10 +103,10 @@ public class TestParserExec extends BaseTest {
 	 */
 	// TODO: port to test framework (can we simplify the Psl grammar?)
 	@Test public void testFailedPredicateExceptionState() throws Exception {
-		String grammar = load("Psl.g4", "UTF-8");
+		String grammar = load(TestParserExec.class.getResource("Psl.g4"), "UTF-8");
 		String found = execParser("Psl.g4", grammar, "PslParser", "PslLexer", "floating_constant", " . 234", false);
 		assertEquals("", found);
-		assertEquals("line 1:6 rule floating_constant DEC:A floating-point constant cannot have internal white space\n", stderrDuringParse);
+		assertEquals("line 1:6 rule floating_constant DEC:A floating-point constant cannot have internal white space\n", stderrDuringParse());
 	}
 
 
@@ -134,6 +135,6 @@ public class TestParserExec extends BaseTest {
 		String input = "2 9 10 3 1 2 3";
 		String found = execParser("Data.g4", grammar, "DataParser", "DataLexer", "file", input, false);
 		assertEquals("6\n", found);
-		assertNull(stderrDuringParse);
+		assertNull(stderrDuringParse());
 	}
 }
