+++ tool/test/org/antlr/v4/test/tool/TestVocabulary.java
@@ -32,6 +32,7 @@ package org.antlr.v4.test.tool;
 import org.antlr.v4.runtime.Token;
 import org.antlr.v4.runtime.Vocabulary;
 import org.antlr.v4.runtime.VocabularyImpl;
+import org.antlr.v4.test.AntlrTestcase;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -39,7 +40,7 @@ import org.junit.Test;
  *
  * @author Sam Harwell
  */
-public class TestVocabulary extends BaseTest {
+public class TestVocabulary extends AntlrTestcase {
 
 	@Test
 	public void testEmptyVocabulary() {
